[fix/flink][taier-ui]fix flink tab saving
